requirements.txt
uvicorn
fastapi
python-telegram-bot==21.2
requests
pytesseract